requirements.txt
fastapi==0.89.1
numpy==1.24.1
pandas==1.5.2
scikit-surprise==1.1.3
scipy==1.10.0
uvicorn==0.20.0